# services/auth/api/api.bm.go
// Code generated by protoc-gen-bm v0.1, DO NOT EDIT.
// source: api.proto

/*
Package api is a generated blademaster stub package.
This code was generated with kratos/tool/protobuf/protoc-gen-bm v0.1.

package 命名使用 {appid}.{version} 的方式, version 形如 v1, v2 ..

It is generated from these files:
	api.proto
*/
package api

import (
	"context"

	bm "github.com/bilibili/kratos/pkg/net/http/blademaster"
	"github.com/bilibili/kratos/pkg/net/http/blademaster/binding"
)
import google_protobuf1 "github.com/golang/protobuf/ptypes/empty"

// to suppressed 'imported but not used warning'
var _ *bm.Context
var _ context.Context
var _ binding.StructValidator

var PathAuthPing = "/auth.service.v1.Auth/Ping"
var PathAuthLogin = "/user/login"
var PathAuthUserSave = "/user/save"

// AuthBMServer is the server API for Auth service.
type AuthBMServer interface {
	Ping(ctx context.Context, req *google_protobuf1.Empty) (resp *google_protobuf1.Empty, err error)

	Login(ctx context.Context, req *UserLoginReq) (resp *UserLoginResp, err error)

	UserSave(ctx context.Context, req *UserSaveReq) (resp *UserSaveResp, err error)
}

var AuthSvc AuthBMServer

func authPing(c *bm.Context) {
	p := new(google_protobuf1.Empty)
	if err := c.BindWith(p, binding.Default(c.Request.Method, c.Request.Header.Get("Content-Type"))); err != nil {
		return
	}
	resp, err := AuthSvc.Ping(c, p)
	c.JSON(resp, err)
}

func authLogin(c *bm.Context) {
	p := new(UserLoginReq)
	if err := c.BindWith(p, binding.Default(c.Request.Method, c.Request.Header.Get("Content-Type"))); err != nil {
		return
	}
	resp, err := AuthSvc.Login(c, p)
	c.JSON(resp, err)
}

func authUserSave(c *bm.Context) {
	p := new(UserSaveReq)
	if err := c.BindWith(p, binding.Default(c.Request.Method, c.Request.Header.Get("Content-Type"))); err != nil {
		return
	}
	resp, err := AuthSvc.UserSave(c, p)
	c.JSON(resp, err)
}

// RegisterAuthBMServer Register the blademaster route
func RegisterAuthBMServer(e *bm.Engine, server AuthBMServer) {
	AuthSvc = server
	e.GET("/auth.service.v1.Auth/Ping", authPing)
	e.POST("/user/login", authLogin)
	e.POST("/user/save", authUserSave)
}