# services/auth/api/api.swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "api.proto",
        "version": "1"
    },
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json",
        "multipart/form-data"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/auth.service.v1.Auth/Ping": {
            "get": {
                "summary": "/auth.service.v1.Auth/Ping",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.google.protobuf.Empty"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "auth.service.v1.Auth"
                ]
            }
        },
        "/user/login": {
            "post": {
                "summary": "/user/login",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.auth.service.v1.UserLoginResp"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/.auth.service.v1.UserLoginReq"
                        }
                    }
                ],
                "tags": [
                    "auth.service.v1.Auth"
                ]
            }
        },
        "/user/save": {
            "post": {
                "summary": "/user/save",
                "responses": {
                    "200": {
                        "description": "A successful response.",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "code": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                },
                                "data": {
                                    "$ref": "#/definitions/.auth.service.v1.UserSaveResp"
                                }
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/.auth.service.v1.UserSaveReq"
                        }
                    }
                ],
                "tags": [
                    "auth.service.v1.Auth"
                ]
            }
        }
    },
    "definitions": {
        ".auth.service.v1.UserLoginReq": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            },
            "required": [
                "username",
                "password"
            ]
        },
        ".auth.service.v1.UserLoginResp": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        ".auth.service.v1.UserSaveReq": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            },
            "required": [
                "username",
                "password"
            ]
        },
        ".auth.service.v1.UserSaveResp": {
            "type": "object",
            "properties": {
                "username": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            },
            "required": [
                "username",
                "password"
            ]
        },
        ".google.protobuf.Empty": {
            "type": "object",
            "properties": {},
            "description": "A generic empty message that you can re-use to avoid defining duplicated\n empty messages in your APIs. A typical example is to use it as the request\n or the response type of an API method. For instance:\n\n     service Foo {\n       rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty);\n     }\n\n The JSON representation for `Empty` is empty JSON object `{}`."
        }
    }
}